restructure tests to enable 100% test coverage
