+++ .taprc
@@ -1,1 +1,1 @@
-check-coverage: false
+check-coverage: true
